requirements.txt
requests
git+https://github.com/rnbr04/epicstore_api
django>=6.0,<6.1
gunicorn>=23,<24
dj-database-url>=3,<4
whitenoise>=6,<7
psycopg[binary]
djangorestframework